refactor(utils): 将 normalizeResourceItem 迁移至 utils/normalize
